Refactor routes so they are split into separate files
Other Changes:
- /users route now has aliases /user and /u, user check moved into user subrouter
- /images route now has aliases /image and /i
- /check_username is now /check/username
- /change_password is now /settings/change_password
- Add middleware for session handler
- Break out other middlewares into their own files
- Minor auto-formatting changes
- Require dotenv only once, at the start. No need to do it twice.
- Update Sentry libraries to 6.19.6. Originally was going to add Sentry to backend, but pulled it for now since I can't figure out how to strip out request bodies with sensitive info (ie. password input) from trace samples at the moment.
