+++ .prettierrc
@@ -1,6 +1,6 @@
 {
     "tabWidth": 4,
-    "printWidth": 100,
+    "printWidth": 120,
     "overrides": [
         {
             "files": "package*.json",
