+++ lib/database-ops.js
@@ -1,4 +1,3 @@
-require("dotenv").config();
 const MongoClient = require("mongodb").MongoClient;
 const ObjectID = require("mongodb").ObjectID;
 const util = require("./util");
@@ -63,8 +62,7 @@ const insertImageIntoDatabase = function (imageEntry, callback) {
                 callback(
                     {
                         status: "error",
-                        message:
-                            "An error occurred when uploading your image. Please try again later.",
+                        message: "An error occurred when uploading your image. Please try again later.",
                     },
                     null
                 );
@@ -189,9 +187,7 @@ const startDatabaseClient = function (callback, options) {
                         })
                         .catch((err) => {
                             if (err.codeName === "IndexNotFound") {
-                                logger.info(
-                                    "expire_index for images collection not found. Creating new expire_index"
-                                );
+                                logger.info("expire_index for images collection not found. Creating new expire_index");
                                 createIndexForImageEntriesCollection(imageEntryCollection);
                                 finishInitialization();
                             } else {
@@ -437,14 +433,12 @@ const findCommentsForImage = function (imageID, callback) {
 };
 
 const findCommentsForUser = function (username, callback) {
-    commentsCollection
-        .find({ username: { $regex: username, $options: "i" } })
-        .toArray(function (err, results) {
-            if (err) {
-                logger.error("There was an error finding comments for username: " + username);
-            }
-            callback(err, results);
-        });
+    commentsCollection.find({ username: { $regex: username, $options: "i" } }).toArray(function (err, results) {
+        if (err) {
+            logger.error("There was an error finding comments for username: " + username);
+        }
+        callback(err, results);
+    });
 };
 
 const findImagesForUser = function (username, callback) {
@@ -457,9 +451,7 @@ const findImagesForUser = function (username, callback) {
         }
         imageEntryCollection.find({ username }).toArray(function (err, results) {
             if (err) {
-                logger.error(
-                    "There was an error finding images for provided username: " + username
-                );
+                logger.error("There was an error finding images for provided username: " + username);
                 callback(err, null);
                 return;
             }
@@ -643,8 +635,7 @@ const writeActionHistoryEntry = function (actionEntry, callback) {
                 logger.error(err);
                 callback(
                     {
-                        message:
-                            "Could not write action history entry to database due to an error.",
+                        message: "Could not write action history entry to database due to an error.",
                     },
                     null
                 );
