+++ lib/middleware/format.js
@@ -1,0 +1,6 @@
+module.exports = (format) => {
+    return (req, res, next) => {
+        req.format = format;
+        next();
+    };
+};
