+++ lib/server.js
@@ -1,8 +1,15 @@
 const express = require("express");
-const route = require("./routes/route");
+const route = require("./route");
 const https = require("https");
 const fs = require("fs");
-const logger = require("./logger").logger;
+const { logger } = require("./logger");
+const bodyParser = require("body-parser");
+const format = require("./middleware/format");
+const { injectUser, sessionsMiddleware } = require("./middleware/session");
+const { sendError } = require("./util/response");
+const { csrfProtection, injectCsrfToken } = require("./middleware/csrf");
+const errorHandler = require("./middleware/error");
+
 
 var app = express();
 var router = express.Router();
@@ -11,7 +18,47 @@ let httpServer, httpsServer;
 app.use(express.static("public"));
 app.set("view engine", "ejs");
 
-route(app, router);
+app.use(bodyParser.json());
+app.use(
+    bodyParser.urlencoded({
+        extended: true,
+    })
+);
+app.set("trust proxy", 1);
+
+app.use(sessionsMiddleware());
+
+app.use(csrfProtection);
+app.use(injectCsrfToken);
+
+router.use("/", require("./route"));
+
+const imagesRouter = require("./route/image");
+router.use("/images", imagesRouter);
+router.use("/image", imagesRouter);
+router.use("/i", imagesRouter);
+
+const usersRouter = require("./route/user");
+router.use("/users", usersRouter);
+router.use("/user", usersRouter);
+router.use("/u", usersRouter);
+
+router.use("/upload", require("./route/upload"));
+
+router.use("/comment", require("./route/comment"));
+
+router.use("/login", require("./route/login"));
+router.use("/register", require("./route/register"));
+router.use("/logout", require("./route/logout"));
+
+router.use("/check", require("./route/check"));
+
+router.use("/settings", require("./route/settings"));
+
+
+app.use("/", router);
+
+app.use(errorHandler);
 
 module.exports.runServer = function (portNumber, callback) {
     httpServer = app.listen(portNumber, () => {
