+++ package.json
@@ -20,8 +20,8 @@
     "watch:client": "webpack -w"
   },
   "dependencies": {
-    "@sentry/browser": "^6.19.4",
-    "@sentry/tracing": "^6.19.4",
+    "@sentry/browser": "^6.19.6",
+    "@sentry/tracing": "^6.19.6",
     "@sentry/webpack-plugin": "^1.18.8",
     "bcrypt": "^5.0.0",
     "body-parser": "^1.18.2",
