+++ views/index.ejs
@@ -1,52 +1,49 @@
 <!DOCTYPE html>
 <html>
+    <head>
+        <meta charset="utf-8" />
+        <title>simpleimage - Simple Image Upload</title>
+        <link rel="stylesheet" href="/assets/css/style.css" />
+        <link rel="stylesheet" href="/assets/css/form-style.css" />
+        <link rel="stylesheet" href="/assets/css/home-style.css" />
+        <script
+            src="https://code.jquery.com/jquery-3.1.1.min.js"
+            integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="
+            crossorigin="anonymous"
+        ></script>
+        <script src="assets/js/home.bundle.js"></script>
+        <script type="text/javascript">
+            // TODO add dedicated config object so that the environment variable can be mapped nicely to a boolean
+            var isLoginRequired = '<%= process.env.LOGIN_TO_UPLOAD === "true" %>';
+        </script>
 
-<head>
-    <meta charset="utf-8">
-    <title>simpleimage - Simple Image Upload</title>
-    <link rel="stylesheet" href="/assets/css/style.css">
-    <link rel="stylesheet" href="/assets/css/form-style.css">
-    <link rel="stylesheet" href="/assets/css/home-style.css">
-    <script src="https://code.jquery.com/jquery-3.1.1.min.js" integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="
-        crossorigin="anonymous"></script>
-    <script src="assets/js/home.bundle.js"></script>
-    <script type="text/javascript">
-        // TODO add dedicated config object so that the environment variable can be mapped nicely to a boolean
-        var isLoginRequired = '<%= process.env.LOGIN_TO_UPLOAD === "true" %>';
-    </script>
-    
-    <%- include("headers") %>
-</head>
+        <%- include("headers") %>
+    </head>
 
-<body>
-    <div id="overlay-backdrop">
-    </div>
-    <%- include("top-nav-view", {user}) %>
-    <div id="contents">
-        <div id="notification-overlay-container">
-        </div>
-        <div id="overlay-container">
-        </div>
-        <div id="upload-center">
-            <div id="upload-area">
-                <div id="upload-preview">
-                </div>
-                <div class="button" id="select-button">
-                    <form action="/upload" id="upload" method="POST" enctype="multipart/form-data">
-                        <input type="file" accept=".gif, .png, .jpg, .jpeg, .bmp" id="select-me" value="">
-                    </form>
-                    <span class="collecticons collecticons-picture head-icon"></span>
-                </div>
-                <div class="button" id="upload-button">
-                    <span class="collecticons collecticons-upload-2 head-icon"></span>
+    <body>
+        <div id="overlay-backdrop"></div>
+        <%- include("top-nav-view", {user}) %>
+        <div id="contents">
+            <div id="notification-overlay-container"></div>
+            <div id="overlay-container"></div>
+            <div id="upload-center">
+                <div id="upload-area">
+                    <div id="upload-preview"></div>
+                    <div class="button" id="select-button">
+                        <form action="/upload" id="upload" method="POST" enctype="multipart/form-data">
+                            <input type="file" accept=".gif, .png, .jpg, .jpeg, .bmp" id="select-me" value="" />
+                        </form>
+                        <span class="collecticons collecticons-picture head-icon"></span>
+                    </div>
+                    <div class="button" id="upload-button">
+                        <span class="collecticons collecticons-upload-2 head-icon"></span>
+                    </div>
                 </div>
             </div>
+            <%- include("footer") %>
         </div>
-        <%- include("footer") %>
-    </div>
-    <script>
-        var csrfToken = "<%= csrfToken %>";
-    </script>
-</body>
-
+        <script>
+            var csrfToken = "<%= csrfToken %>";
+        </script>
+    </body>
 </html>
