+++ views/settings-view.ejs
@@ -32,11 +32,12 @@
         };
 
         var submitPasswordReset = function() {
-            var form = $("form[id='change-password-form']");
-            var req = new XMLHttpRequest();
+            const form = $("form[id='change-password-form']");
+            const action = form.attr("action");
+            const req = new XMLHttpRequest();
 
             req.onload = onPasswordResetSubmitted;
-            req.open("post", "/change_password");
+            req.open("post", action);
             req.setRequestHeader("Content-Type", "application/x-www-form-urlencoded");
             req.send(form.serialize());
         }
@@ -70,7 +71,7 @@
             </div>
             <div class="settings-view">
                 <span>Change Password</span><br/>
-                <form action="/change_password" method="POST" id="change-password-form" class="input-form">
+                <form action="/settings/change_password" method="POST" id="change-password-form" class="input-form">
                     <span>Old Password</span>
                     <input type="password" name="oldPassword"></input><br />
                     <span>New Password</span>
